% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_argos.R
\name{argos_filter}
\alias{argos_filter}
\title{Argos Filter for Position Data}
\usage{
argos_filter(
  posdata,
  light_data_calibration,
  logger_colony_info,
  logger_filter
)
}
\arguments{
\item{posdata}{A data frame containing position data with columns \code{lat_smooth2}, \code{lon_smooth2}, \code{date_time}, \code{type}, and \code{eqfilter}.}

\item{light_data_calibration}{A list containing light data calibration parameters.}

\item{logger_colony_info}{A list containing logger colony information, including \code{col_lat} and \code{col_lon}.}

\item{logger_filter}{A list containing logger filter parameters, including \code{speed}.}
}
\value{
A filtered data frame with positions classified as "removed" or "end_location" by the Argos SDA filter removed, except during equinox periods.
}
\description{
Filters position data using the Argos SDA filter, removing positions classified as "removed" or "end_location", while considering equinox periods.
}
\concept{filtering}
