% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_aut_midnight_sun_removal.R
\name{aut_midnight_sun_removal}
\alias{aut_midnight_sun_removal}
\title{Remove false positions during breeding or midnight sun periods}
\usage{
aut_midnight_sun_removal(df)
}
\arguments{
\item{df}{data.frame where  'tFirst' and 'tSecond' (both date_time) must be part of the first eight columns}
}
\value{
A data.frame where non-consequative positions have been removed from the head and tail of the original data.frame
}
\description{
At the head or tail of a yearly track (e.g breedingY to breedingY+1) : removes the first or last positions that are more than 25 hours apart in time from the other positions.
Datagaps of >25 hours in the midst of the track won't be assessed or filtered.
}
