% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_bounding_box.R
\name{bounding_box_filter}
\alias{bounding_box_filter}
\title{Bounding Box Filter for Position Data}
\usage{
bounding_box_filter(posdata, light_data_calibration, logger_filter)
}
\arguments{
\item{posdata}{A data frame containing position data with columns \code{lon}, \code{lat}, and \code{tFirst}.}

\item{light_data_calibration}{A list containing light data calibration parameters.}

\item{logger_filter}{A list containing logger filter parameters, including default boundary box limits.}
}
\value{
A filtered data frame with positions outside the bounding box removed.
}
\description{
Filters position data based on a defined bounding box, removing positions outside the specified latitude and longitude limits.
}
\concept{filtering}
