% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_calibration_to_wb.R
\name{calibration_to_wb}
\alias{calibration_to_wb}
\title{Write Calibration Data to Excel Workbook}
\usage{
calibration_to_wb(
  all_calibration,
  calibration_output_dir,
  calibration_filename = paste0("calibration_", Sys.Date(), ".xlsx")
)
}
\arguments{
\item{all_calibration}{Data frame containing calibration data for all loggers.}

\item{calibration_output_dir}{Directory to save the calibration Excel workbook.}

\item{calibration_filename}{Name of the calibration Excel file. Defaults to "calibration_<current_date>.xlsx".}
}
\value{
None. Saves the calibration data to an Excel workbook at the specified file path.
}
\description{
Writes the provided calibration data frame to an Excel workbook with formatting.
}
\concept{function}
