% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_compare_sun_angle.R
\name{compare_sun_angle}
\alias{compare_sun_angle}
\title{Compare Sun Angles to Optimize Position Estimation}
\usage{
compare_sun_angle(prev_posdata_export, new_posdata_export, type)
}
\arguments{
\item{prev_posdata_export}{A data frame containing previous position data export with columns \code{eqfilter}, \code{lat}, \code{type}, and \code{date_time}.}

\item{new_posdata_export}{A data frame containing new position data export with columns \code{tFirst}, \code{tSecond}, \code{type}.}

\item{type}{A string indicating the type of sun angle sequence to use. Options are \code{"general"}, \code{"summer"}, or \code{"winter"}.}
}
\value{
A list containing the optimal sun angles for the start and end of the track: \code{sun_angle_start} and \code{sun_angle_end}.
}
\description{
This function compares sun angles from previous and new position data exports to optimize position estimation based on latitude differences.
}
\keyword{internal}
