% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_filter_file.R
\name{create_filter_file}
\alias{create_filter_file}
\title{Create Filter File}
\usage{
create_filter_file(filepath, species = c())
}
\arguments{
\item{filepath}{Filepath to save the filter settings Excel file.}

\item{species}{Vector of species names to include in the filter file. If empty, creates an empty template with default settings for black-legged kittiwake.}
}
\value{
None. The function creates an Excel file at the specified filepath.
}
\description{
Create an example filter settings Excel file for specified species.
The file will contain default filter settings for each species.
See `read_filter_file` for reading the filter file.
}
\concept{custom_filters}
