% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_date_conversion.r
\name{date_conversion}
\alias{date_conversion}
\title{convert dates for seatrackGLS}
\usage{
date_conversion(date, tz = "UTC")
}
\arguments{
\item{date}{Date to be converted}
}
\value{
Converted date in format "yyyy-mm-dd UTC"
}
\description{
convert dates for seatrackGLS
}
