% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_datetime_conversion.r
\name{datetime_conversion}
\alias{datetime_conversion}
\title{convert date time for seatrackGLS}
\usage{
datetime_conversion(datetime, tz = "UTC")
}
\arguments{
\item{date}{Date and time to be converted}
}
\value{
Converted date in format "yyyy-mm-dd hh:mm:ss UTC"
}
\description{
convert date time for seatrackGLS
}
\keyword{internal}
