% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_equinox.R
\name{equinox_filter}
\alias{equinox_filter}
\title{Equinox Filter for Position Data}
\usage{
equinox_filter(
  posdata,
  posdata_lat,
  light_data_calibration,
  logger_colony_info
)
}
\arguments{
\item{posdata}{A data frame containing position data with columns \code{date_time}.}

\item{posdata_lat}{A numeric vector of latitudes corresponding to the position data.}

\item{light_data_calibration}{A list containing light data calibration parameters, including equinox start and end dates.}

\item{logger_colony_info}{A list containing logger colony information, including \code{col_lat}.}
}
\value{
A data frame with an additional column \code{eqfilter} indicating equinox effect presence (TRUE = no equinox effect, FALSE = equinox effect).
}
\description{
Applies an equinox filter to position data based on calibrated light data.
}
\concept{filtering}
