% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equinox_table}
\alias{equinox_table}
\title{Equinox table}
\format{
## `equinox_table`
A data frame with 4,669 rows and 10 columns:
\describe{
  \item{latitude}{Latitude}
  \item{sun_angle}{Angle of sun}
}
}
\usage{
equinox_table
}
\description{
Vegard's magical table of equinoxes
}
\keyword{datasets}
