% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_metadata}
\alias{example_metadata}
\alias{example_colony_info}
\alias{example_extra_metadata}
\title{Example metadata for GLS loggers
A dataset containing example metadata for GLS loggers.}
\format{
An object of class \code{data.frame} with 1 rows and 6 columns.

An object of class \code{data.frame} with 1 rows and 3 columns.

An object of class \code{data.frame} with 1 rows and 5 columns.
}
\usage{
example_metadata

example_colony_info

example_extra_metadata
}
\description{
Example metadata for GLS loggers
A dataset containing example metadata for GLS loggers.

Example colony information for GLS loggers
A dataset containing example example colony information for GLS loggers.

Example extra metadata for GLS loggers
A dataset containing example extra metadata for GLS loggers.
}
\concept{data}
\keyword{datasets}
