% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_filter_plot_export.R
\name{export_filter_plot}
\alias{export_filter_plot}
\title{Wrapper to export filter plots}
\usage{
export_filter_plot(
  call = {
 },
  show_filter_plots = TRUE,
  plot_name = "",
  logger_id_year = "",
  plotting_dir = NULL
)
}
\arguments{
\item{call}{The code to be evaluated, typically a function call that generates a plot.}

\item{show_filter_plots}{Logical indicating whether to show and save filter plots.}

\item{plot_name}{A string specifying the name of the plot file (without extension).}

\item{logger_id_year}{A string to identify the logger and year, used in the plot filename.}

\item{filter_plots_dir}{The directory where the plot files should be saved. If NULL, plots are not saved to files.}
}
\value{
The result of evaluating the provided code call.
}
\description{
This function wraps around a code call to export plots used in filtering steps.
If \code{show_filter_plots} is FALSE, the code is simply evaluated without plotting
Otherwise, if \code{show_filter_plots} is TRUE, the code is evaluated and the resulting plot is saved to a PNG file in the specified \code{filter_plots_dir}.
}
\keyword{internal}
