% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_breeding_month_seq}
\alias{get_breeding_month_seq}
\title{Generate a sequence of breeding months}
\usage{
get_breeding_month_seq(start_month = 6, end_month = 8)
}
\arguments{
\item{start_month}{An integer representing the starting month of the breeding period (1-12).}

\item{end_month}{An integer representing the ending month of the breeding period (1-12).}
}
\value{
A vector of integers representing the months in the breeding period.
}
\description{
This function generates a sequence of months representing the breeding period,
taking into account cases where the breeding period spans the end of the year.
}
