% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_get_seatrack_metadata.R
\name{get_seatrack_metadata}
\alias{get_seatrack_metadata}
\title{Extract SEATRACK metadata to inform processing of light data into positions}
\usage{
get_seatrack_metadata(Colony, Species, Analyzer)
}
\arguments{
\item{Colony}{name of colony, must correspond to the colony name in SEATRACK database. Required}

\item{Species}{name of species, must correspond to the species name in SEATRACK database. Required}

\item{Analyzer}{name of analyzer in order to keep some track of history. Optional}
}
\value{
A data.frame with metadata extracted from the database. 'Sun' columns left blank.
}
\description{
Require login to the SEATRACK database. FUNCTION TO MOVE TO ANOTHER PACKAGE
}
