% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_filter_file.R
\name{GLSfilterList}
\alias{GLSfilterList}
\title{GLSfilterList}
\description{
A convenience class for storing a list of GLSsettings objects and retrieving settings based on logger/species/colony.
}
\seealso{
Other classes: 
\code{\link{GLSsettings}}
}
\concept{classes}
\concept{custom_filters}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{filter_list}}{List of GLSsettings objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GLSfilterList-new}{\code{GLSfilterList$new()}}
\item \href{#method-GLSfilterList-get_settings_from_list}{\code{GLSfilterList$get_settings_from_list()}}
\item \href{#method-GLSfilterList-clone}{\code{GLSfilterList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GLSfilterList-new"></a>}}
\if{latex}{\out{\hypertarget{method-GLSfilterList-new}{}}}
\subsection{Method \code{new()}}{
Create a new GLSfilterList object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GLSfilterList$new(filter_list = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filter_list}}{A list of GLSsettings objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GLSfilterList-get_settings_from_list"></a>}}
\if{latex}{\out{\hypertarget{method-GLSfilterList-get_settings_from_list}{}}}
\subsection{Method \code{get_settings_from_list()}}{
Get Settings from settings_list based on logger ID, species, or colony.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GLSfilterList$get_settings_from_list(
  species = NULL,
  colony = NULL,
  logger_id = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{species}}{Optional species name to match.}

\item{\code{colony}}{Optional colony name to match.}

\item{\code{logger_id}}{Optional logger ID to match.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of filter settings matching the provided identifiers. If no match is found, returns default settings for the species.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GLSfilterList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GLSfilterList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GLSfilterList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
