% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_settings.R
\name{GLSFilterSettingsList}
\alias{GLSFilterSettingsList}
\title{GLSFilterSettingsList}
\description{
A convenience class for storing a list of \code{GLSsettings} objects and retrieving settings based on logger/species/colony.
}
\seealso{
Other classes: 
\code{\link{GLSsettings}}
}
\concept{classes}
\concept{custom_filters}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{filter_list}}{List of GLSsettings objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GLSFilterSettingsList-new}{\code{GLSFilterSettingsList$new()}}
\item \href{#method-GLSFilterSettingsList-get_settings_from_list}{\code{GLSFilterSettingsList$get_settings_from_list()}}
\item \href{#method-GLSFilterSettingsList-clone}{\code{GLSFilterSettingsList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GLSFilterSettingsList-new"></a>}}
\if{latex}{\out{\hypertarget{method-GLSFilterSettingsList-new}{}}}
\subsection{Method \code{new()}}{
Create a new GLSFilterSettingsList object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GLSFilterSettingsList$new(filter_list = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filter_list}}{A list of GLSsettings objects.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GLSFilterSettingsList-get_settings_from_list"></a>}}
\if{latex}{\out{\hypertarget{method-GLSFilterSettingsList-get_settings_from_list}{}}}
\subsection{Method \code{get_settings_from_list()}}{
Get Settings from settings_list based on logger ID, species, or colony.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GLSFilterSettingsList$get_settings_from_list(
  species = NULL,
  colony = NULL,
  logger_id = NULL,
  years_tracked = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{species}}{Optional species name to match.}

\item{\code{colony}}{Optional colony name to match.}

\item{\code{logger_id}}{Optional logger ID to match.}

\item{\code{years_tracked}}{Optional year tracked to match}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of filter settings matching the provided identifiers. If no match is found, returns default settings for the species.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GLSFilterSettingsList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GLSFilterSettingsList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GLSFilterSettingsList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
