% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_filter_file.R
\name{GLSsettings}
\alias{GLSsettings}
\title{GLSsettings Class}
\description{
A convenience class for storing settings lists alongside identifiers to which logger/species/colony they belong
}
\seealso{
Other classes: 
\code{\link{GLSfilterList}}
}
\concept{classes}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{species}}{Species identifier}

\item{\code{colony}}{Colony identifier}

\item{\code{logger_id}}{Logger ID identifier}

\item{\code{settings}}{A list containing filter settings for the logger/species/colony.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GLSsettings-new}{\code{GLSsettings$new()}}
\item \href{#method-GLSsettings-check_settings_for}{\code{GLSsettings$check_settings_for()}}
\item \href{#method-GLSsettings-clone}{\code{GLSsettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GLSsettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-GLSsettings-new}{}}}
\subsection{Method \code{new()}}{
Create a new GLSsettings object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GLSsettings$new(
  logger_id = NULL,
  species = NULL,
  colony = NULL,
  settings = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{logger_id}}{Optional logger ID to associate with these settings.}

\item{\code{species}}{Optional species name to associate with these settings.}

\item{\code{colony}}{Optional colony name to associate with these settings.}

\item{\code{settings}}{Optional list of filter settings. If NULL, defaults to settings for the provided species from `seatrackRgls::seatrack_settings_list`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GLSsettings-check_settings_for"></a>}}
\if{latex}{\out{\hypertarget{method-GLSsettings-check_settings_for}{}}}
\subsection{Method \code{check_settings_for()}}{
Check if these settings are for a particular combination of logger/species/colony
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GLSsettings$check_settings_for(species = NULL, logger_id = NULL, colony = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{species}}{Optional species name to match.}

\item{\code{logger_id}}{Optional logger ID to match.}

\item{\code{colony}}{Optional colony name to match.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if the settings match the provided identifiers, FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GLSsettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GLSsettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GLSsettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
