% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLight_i.preSelection_bugfree.R
\name{i.preSelection_bugfree}
\alias{i.preSelection_bugfree}
\title{Function from GeoLight package. Property of GeoLight and Lisovski 2012.}
\usage{
i.preSelection_bugfree(datetime, light, LightThreshold)
}
\description{
Property of GeoLight and Lisovski 2012. Small bug fixed in an outdated date-formatting function from baseR
}
\keyword{internal}
