% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLight_i.twilightEvents.R
\name{i.twilightEvents}
\alias{i.twilightEvents}
\title{Function from GeoLight package. Property of GeoLight and Lisovski 2012.}
\usage{
i.twilightEvents(datetime, light, LightThreshold)
}
\description{
Property of GeoLight and Lisovski 2012. Small bug fixed in an outdated date-formatting function from baseR
}
\keyword{internal}
