% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_land_mask.R
\name{land_mask}
\alias{land_mask}
\title{Land mask}
\usage{
land_mask(lon, lat, coast_to_land, coast_to_sea, eqfilter)
}
\arguments{
\item{lon}{longitude in decimal degrees}

\item{lat}{latitude in decimal degrees}

\item{coast_to_land}{locations filtered if X km inland from coastline}

\item{coast_to_sea}{locations filtered if X km offshore from the coastline}

\item{eqfilter}{if equinox, locations are not filtered}
}
\value{
A data frame with longitude and latitude filtered.
}
\description{
The function measures the distance between any location and to a coastline, using the package ‘distancetocoast’ and a high-resolution coastline from Natural Earth ("10").
Land and ocean location can  be removed surpassing a specified distance in km.
To avoid this function, apply very far (e.g '10000' km) or infinite ('Inf') distance.
}
