% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_load_light_data.R
\name{limit_light_data}
\alias{limit_light_data}
\title{Limit light data to calibration time windows}
\usage{
limit_light_data(light_data, logger_calibration_data)
}
\arguments{
\item{light_data}{Data frame containing light data with a 'dtime' column.}

\item{logger_calibration_data}{Data frame containing calibration data with 'start_datetime' and 'end_datetime' columns.}
}
\value{
A list of data frames, each containing light data limited to the corresponding calibration time window.
}
\description{
Limit light data to calibration time windows
}
\keyword{internal}
