% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_loess.R
\name{loess_filter}
\alias{loess_filter}
\title{Loess Filter for Position Data}
\usage{
loess_filter(posdata, logger_filter)
}
\arguments{
\item{posdata}{A data frame containing position data with columns \code{tFirst}, \code{tSecond}, and \code{type}.}

\item{logger_filter}{A list containing logger filter parameters, including `loess_filter_k}
}
\value{
A data frame with loess-filtered positions if applicable; otherwise, the original data frame.
}
\description{
Applies a loess filter to position data if the track length exceeds 45 days.
}
\concept{filtering}
