% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_noonfilter.R
\name{noonfilter}
\alias{noonfilter}
\title{Remove twilights that form subsequent outliers in time of noon and midnight.}
\usage{
noonfilter(df, show_plot)
}
\arguments{
\item{df}{Input data with twilight times in GMT shown as 'tfirst' and 'tsecond' (yyyy-mm-dd hh:mm:ss UTC) and 'type', with being 1 (tfirst is sunrise) or 2 (tfirst is sunset).}

\item{show_plot}{if TRUE, plot is produced with timing of noon/midnight and affected twilights highlighted}
}
\value{
A filtered data.frame in the same format.
}
\description{
This function performs the following steps:
\enumerate{
  \item Calculate noon/midnight based on twilight.
  \item make noon/midnight circular.
  \item split data set into day and night
  \item loess predictions of noon/midnight (after excluding twilights with extreme variation from predictions)
  \item fill data gaps in predictions with baytrends::fillMissing()
  \item find the minimum difference between a noon/midnight directly calculated from twilights and any predicted noon/midnight 4 days ahead or after candidate noon/midnight.
  \item threshold for removing twilight is when difference between noon/midnight vs prediction overreach 0.4h, but under under 12h
  }
}
