% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_plot_a_map.R
\name{plot_a_map}
\alias{plot_a_map}
\title{Produce a Quick Map}
\usage{
plot_a_map(df)
}
\arguments{
\item{df}{data frame with lat, lon, eqfilter, date_time, col_lat, col_lon, colony}
}
\value{
Plotted map.
}
\description{
Produce a Quick Map
}
