% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_folder_wrappers.R
\name{process_positions}
\alias{process_positions}
\title{Process Positions from a Folder}
\usage{
process_positions(
  import_directory,
  calibration_data,
  all_colony_info,
  output_directory,
  extra_metadata = NULL
)
}
\arguments{
\item{import_directory}{Directory containing raw light data files.}

\item{calibration_data}{Data frame containing calibration data for loggers.}

\item{all_colony_info}{Data frame containing colony information.}

\item{output_directory}{Directory to save processed position outputs.}

\item{extra_metadata}{Optional data frame containing extra metadata for loggers.}
}
\description{
Process Positions from a Folder
}
