% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_seatrack_database_format.R
\name{seatrack_database_format}
\alias{seatrack_database_format}
\title{Reformat to SEATRACK database format (outdated per Oct, 2025)}
\usage{
seatrack_database_format(df)
}
\arguments{
\item{df}{seatrackGLS dataset}
}
\value{
dataset in database format
}
\description{
Reformat to SEATRACK database format (outdated per Oct, 2025)
}
