% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_seatrack_database_format_newformat.R
\name{seatrack_database_format_newDBformat}
\alias{seatrack_database_format_newDBformat}
\title{Reformat to SEATRACK database format (NEW per Oct, 2025)}
\usage{
seatrack_database_format_newDBformat(df, species, colony)
}
\arguments{
\item{df}{seatrackGLS dataset}
}
\value{
dataset in database format
}
\description{
Reformat to SEATRACK database format (NEW per Oct, 2025)
}
