% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seatrack_settings_list}
\alias{seatrack_settings_list}
\title{Seatrack filter settings}
\format{
An object of class \code{list} of length 19.
}
\usage{
seatrack_settings_list
}
\description{
Default seatrack filter settings
}
\keyword{datasets}
