% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_aut_processing.R
\name{seatrackGLS}
\alias{seatrackGLS}
\title{Estimate and Refine Locations From Light Data Recorded by Geolocators.}
\usage{
seatrackGLS(
  luxfile,
  save_sun_plots.dir,
  save_filter_plots.dir,
  speed,
  boundary.box,
  coast_to_land,
  coast_to_sea,
  loess_filter_k,
  months_breeding,
  man_equinox_periods,
  aut_equinox_periods,
  noonfilter,
  daylengthfilter,
  midnightsun_removal,
  add_summer_pos,
  add_winter_pos,
  split_years,
  year_tracked,
  sun_angle_start,
  sun_angle_end,
  light_threshold,
  logger_id,
  logger_model,
  ring_number,
  species,
  colony,
  col_lon,
  col_lat,
  date_deployed,
  date_retrieved,
  age_deployed
)
}
\arguments{
\item{luxfile}{Input data containing light readings.}

\item{save_sun_plots.dir}{Directory for where to save plots for sun angle calibrations. If NULL, no plots produced.}

\item{save_filter_plots.dir}{Directory for where to save plots fof filter performance. If NULL, no plots produced.}

\item{speed}{Maximum expected movement rate as km/h, sustained between two locations with ~12 hours in between. Second location in a pair will be removed.}

\item{boundary.box}{Distribution limits in decimal degrees (c(West,East,South,North)) PS! lons as 0-360 deg if crossing the 180 meridian (otherwise -180 to 180)}

\item{coast_to_land}{Distance from coastline locations are allowed inland before removed. 'Inf' if no limit should be applied}

\item{coast_to_sea}{Distance from coastline locations are allowed offshore before removed. 'Inf' if no limit should be applied}

\item{loess_filter_k}{N interquartile ranges used to filter outliers in twilight timing using the a loess function}

\item{months_breeding}{Expected months with regular presence at breeding location}

\item{man_equinox_periods}{Set start and end day of a spring and autumn equinox period where latitudes are not reliable. Previously settings in SEATRACK:'c("02-20","04-03","09-08","10-20")'}

\item{aut_equinox_periods}{If TRUE - start and end point of equinox periods is informed by birds' estimated latitude before and after equinox, plus the calibrated sun angle value.}

\item{noonfilter}{If TRUE - remove twilights that form outliers in timing of noon or midnight.}

\item{daylengthfilter}{If TRUE - remove twilights that form outliers in length of day or night.}

\item{midnightsun_removal}{If TRUE - helps removing false location estimates while breeding under the midnight sun}

\item{add_summer_pos}{If TRUE - add locations to dates where data is missing in the main manually calibrated track using a high threshold of light}

\item{add_winter_pos}{If TRUE - add locations to dates where data is missing in the main manually calibrated track using a minimum threshold of light}

\item{split_years}{Define day/month where to end and start (split) tracking year. E.g. 'c("05-31","06-01")'}

\item{year_tracked}{e.g. June 2015 to May 2016 = "2015_16", January 2016 to December 2016 = "2016_16"}

\item{sun_angle_start}{if NA, default is -3.5. Analyser to add manually calibrated value after an initial run of the function}

\item{sun_angle_end}{filled if compensating for a change in light sensitivity (default is NA)}

\item{light_threshold}{Reference for main manually calibrated track.}

\item{logger_id}{default NA, information can be added for convenience}

\item{logger_model}{default NA, specifying model name can inform algorithm on certain exceptions}

\item{ring_number}{default NA, information can be added for convenience}

\item{species}{default NA, information can be added for convenience}

\item{colony}{default NA, information can be added for convenience}

\item{col_lon}{longitude where bird was instrumented informs the algorithm}

\item{col_lat}{latitude where bird was instrumented informs the algorithm}

\item{date_deployed}{deployment date set start point}

\item{date_retrieved}{retrieval date set end point}

\item{age_deployed}{default NA, information can be added for convenience}
}
\value{
A data frame with raw and smoothed locations, twilight time, threshold and sun angle used to estimate locations, as well as some convenient info about logger and individual.
}
\description{
Light data must be in the format of a '.lig' (BAS/Biotrack/Lotek) or '.lux' (Migrate Technology) file.
Package contains generic settings used in SEATRACK for 16 species.
Procedure use the threshold method (Lisovski et al.2012), and several functions used are from the GeoLight package
Package currently fit to low (1-12 lux, 1-10 bits) and intermediate light thresholds (300 lux, 50 bits), mainly in terms of calibration.
Package will be expanded to adapt itself to any chosen light-level value.
Data needs to be calibrated before the final run in order to produce reliable latitudes.
Calibration values assigned to each tracked non breeding season - see documentation
}
\details{
This function performs the following steps:
\enumerate{
  \item Read in light data and limit to one tracking year, user define where to split between tracking years. At year deployed, 'split_years' is ignored as starting point and first year tracked automatically starts with deployment date. At year retrieved, endpoint in 'split_years' is ignored, and last year tracked ends automatically at date of retrieval.
  \item Estimates twilight times based on a generic threshold of light (see documentation).
  \item Cleans up twilight data, informed by expected movement rate, expected breeding period and breeding location
  \item Estimate latitude and longitude from timing of twilight. If no calibration value given ('sun_angle_start'), assumes 'sun_angle_start' = '-3.5'
  \item Define equinox periods (manually or automatically)
  \item Filter positions based on movement rates and boundaries. Positional filters are limited to longitudes during equinox.
  \item Can produce filter and calibration plots (calibration plots without land mask applied)
  \item Double smoothing of locations
  \item Adds 'winter' and 'summer' locations to the track when the main process miss locations, by automatically running the same process with high and minimum light thresholds.
  \item Final output is a table of filter performance, location table with corresponding twilight, and a twilight table before positional filters were applied.
}
}
