% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_speed.R
\name{speed_filter}
\alias{speed_filter}
\title{Speed Filter for Position Data}
\usage{
speed_filter(posdata, speed)
}
\arguments{
\item{posdata}{A data frame containing position data with columns \code{lat_smooth2}, \code{lon_smooth2}, \code{tFirst}, and \code{eqfilter}.}

\item{speed}{Maximum allowable speed in km/h.}
}
\value{
A filtered data frame with positions exceeding speed limits or showing abnormal longitude changes during equinox periods removed.
}
\description{
Filters position data based on travel speed and longitude changes during equinox periods.
}
\concept{filtering}
