% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_summer_locations.R
\name{summer_locations}
\alias{summer_locations}
\title{Estimate and Filter Summer Locations from GLS Data}
\usage{
summer_locations(
  df,
  luxfile,
  speed,
  boundary.box,
  loess_filter_k,
  months_breeding,
  species,
  midnightsun_removal,
  split_years,
  year_tracked,
  logger_id,
  date_deployed,
  date_retrieved,
  man_equinox_periods,
  aut_equinox_periods
)
}
\arguments{
\item{df}{Already processed and manually calibrated data (main).}

\item{luxfile}{Input data containing light readings.}

\item{speed}{Maximum expected movement rate as km/h, sustained between two locations with ~12 hours in between. Second location in a pair will be removed.}

\item{boundary.box}{Distribution limits in decimal degrees (c(West,East,South,North)) PS! lons as 0-360 deg if crossing the 180 meridian (otherwise -180 to 180)}

\item{loess_filter_k}{N interquartile ranges used to filter outliers in twilight timing using the a loess function.}

\item{months_breeding}{Expected months with regular presence at breeding location.}

\item{species}{default NA, information can be added for convenience.}

\item{midnightsun_removal}{If TRUE - helps removing false location estimates while breeding under the midnight sun.}

\item{split_years}{Define day/month where to end and start (split) tracking year. E.g. 'c("05-31","06-01")'.}

\item{year_tracked}{e.g. June 2015 to May 2016 = "2015_16", January 2016 to December 2016 = "2016_16".}

\item{logger_id}{default NA, information can be added for convenience}

\item{date_deployed}{deployment date set start point}

\item{date_retrieved}{retrieval date set end point}

\item{man_equinox_periods}{Set start and end day of a spring and autumn equinox period where latitudes are not reliable. Previously settings in SEATRACK:'c("02-20","04-03","09-08","10-20")'}

\item{aut_equinox_periods}{If TRUE - start and end point of equinox periods is informed by birds' estimated latitude before and after equinox, plus the calibrated sun angle value.}
}
\value{
A data frame with refined summer locations added to the main manually calibrated track.
}
\description{
This function performs the following steps:
\enumerate{
  \item Estimates twilight times based on a threshold of 50 (for BAS, Biotrack, Lotek), or 300 units (for Migrate Technology).
  \item Cleans up the data using an initial sun angle of 0 degrees.
  \item Identifies the sun angle that best matches the main manually calibrated track by comparing tracks calculated with various sun angles to similar dates in the main manually calibrated track and selecting the most similar one.
  \item Recalculates the track using the most corresponding sun angle.
  \item Add locations to dates where data is missing in the main manually calibrated track.
}
}
