% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_twilight_cleanup.R
\name{twilight_cleanup}
\alias{twilight_cleanup}
\title{Estimate and Refine Locations From Light Data Recorded by Geolocators.}
\usage{
twilight_cleanup(
  df,
  breedingloc_lon,
  breedingloc_lat,
  months_breeding,
  species,
  show_plot,
  sun_angle_start,
  sun_angle_end,
  show_filter_plots = FALSE
)
}
\arguments{
\item{df}{data.frame with 'tFirst', 'tSecond', 'type','sun'.}

\item{breedingloc_lon}{longitude where bird was instrumented informs the algorithm}

\item{breedingloc_lat}{latitude where bird was instrumented informs the algorithm}

\item{months_breeding}{Expected months with regular presence at breeding location}

\item{species}{default NA, information can be added for convenience}

\item{show_plot}{TRUE/FALSE default NA, information can be added for convenience}

\item{sun_angle_start}{if NA, default is -3.5. Analyser to add manually calibrated value after an initial run of the function}

\item{sun_angle_end}{filled if compensating for a change in light sensitivity (default is NA)}
}
\value{
A data frame with raw and smoothed locations, twilight time, threshold and sun angle used to estimate locations, as well as some convenient info about logger and individual.
}
\description{
Filter that remove false days and nights caused by recording shading or artificial light patterns that produce errornous twilights,
The filter is made to identify dates with an impossible number of twilights.
Normally, there is one sunset and one sunrise to a date, but to not affect long-range migration,
the filter allow dates with two sunsets or two sunrises as long as sunsets or sunrises occur more than 22 hours apart from each other.
The function then predict timing of twilights to be used for selecting the most likely set of twilights within these dates.
It use predicted times to avoid referring to singular datapoints that very well can be outliers.
}
\details{
This function performs the following steps:
\enumerate{
\item find dates with too many twilights (1 sunset and 1 sunrise per 22 hours)
\item Build predictions of time of sunset and sunrise by keeping unlikely times out of predictions with loess filter and std deviation.
\item Standard deviation: calculate SD every 5th day to keep points with SD> 60mins out when making predictions
\item Fill predictions for dates with no light with twilights that makes up a daylength of 0 hours.
\item Retain the candidate twilight that is closest to a predicted twilight, remove the others
}
}
\concept{filtering}
