% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilight_estimation.R
\name{twilight_estimation}
\alias{twilight_estimation}
\title{Estimate twilight times from light data}
\usage{
twilight_estimation(
  light_data,
  light_data_calibration,
  show_filter_plots = FALSE
)
}
\arguments{
\item{light_data}{A dataframe containing light data with columns 'dtime' (datetime) and 'lux' (light intensity).}

\item{light_data_calibration}{A list containing calibration parameters, including 'light_threshold' and 'logger_model'.}

\item{show_filter_plots}{A logical value indicating whether to display plots of the filtering process. Default is FALSE.}
}
\value{
a dataframe where the first two columns indicate timing of twilight events and a column type, where type == 1 indicates that tFirst is sunrise and type == 2 indicates that tFirst is sunset.
}
\description{
#' This function estimates twilight times (sunrise and sunset) from light data using the `twilightCalc_bugfree` function.
It also provides an option to visualize the filtering process through plots.
}
