% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoLight_twilightCalc_bugfree.R
\name{twilightCalc_bugfree}
\alias{twilightCalc_bugfree}
\title{Function from GeoLight package. Property of GeoLight and Lisovski 2012.}
\usage{
twilightCalc_bugfree(
  datetime,
  light,
  LightThreshold = TRUE,
  preSelection = TRUE,
  maxLight = NULL,
  ask = TRUE,
  nsee = 500,
  allTwilights = FALSE
)
}
\description{
Property of GeoLight and Lisovski 2012. Small bug fixed in an outdated date-formatting function from baseR
}
\keyword{internal}
