% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equinox_table}
\alias{equinox_table}
\title{Equinox table}
\format{
An object of class \code{data.frame} with 4669 rows and 10 columns.
}
\usage{
equinox_table
}
\description{
Vegard's magical table of equinoxes
\describe{
\item{latitude}{Latitude}
\item{sun_angle}{Angle of sun}
}
}
\concept{data}
\keyword{datasets}
