% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_folder_wrappers.R
\name{prepare_calibration}
\alias{prepare_calibration}
\title{Prepare Calibration Data from a Folder}
\usage{
prepare_calibration(
  import_directory,
  metadata,
  all_colony_info,
  output_directory,
  export_calibration_template = TRUE
)
}
\arguments{
\item{import_directory}{Directory containing raw light data files.}

\item{metadata}{Data frame containing simple metadata for loggers.}

\item{all_colony_info}{Data frame containing colony information.}

\item{output_directory}{Directory to save processed position outputs.}

\item{export_calibration_template}{A logical indicating whether to export a calibration template. Defaults to TRUE.}
}
\description{
Prepare Calibration Data from a Folder
}
