% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_folder.R
\name{process_folder}
\alias{process_folder}
\title{Process all lightposition files in a folder}
\usage{
process_folder(
  import_directory,
  calibration_data,
  filter_list = seatrack_settings_list,
  all_colony_info,
  extra_metadata
)
}
\description{
WIP
}
