% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_logger_year.R
\name{process_logger_year}
\alias{process_logger_year}
\title{Process GLS data for a single logger and year}
\usage{
process_logger_year(
  logger_id,
  year,
  import_directory,
  calibration_data,
  all_colony_info,
  filter_list = seatrack_settings_list,
  extra_metadata = NULL,
  show_filter_plots = FALSE,
  export_maps = TRUE,
  plotting_dir = NULL,
  output_dir = NULL,
  calibration_mode = TRUE,
  split_years = "06-01",
  analyzer = ""
)
}
\arguments{
\item{logger_id}{The ID of the logger to process.}

\item{year}{The END year to process for the logger.}

\item{import_directory}{The directory containing the light data files. File are expected to be named in the format 'logger_id'_`end_year`.}

\item{calibration_data}{A data frame containing calibration data for all loggers or a string providing a filepath to read this data from. This can be from an excel file, CSV file or a directory containing multiple calibration files.
In calibration mode, this dataframe can consist of a single row per logger/year combination.
In calibration mode, the minimum required columns are `logger_id`, `species`, `colony`, `date_deployed` and `date_retrieved`.
If not in calibration mode, the data frame is expected to be in the format output by running this function in calibration mode.}

\item{all_colony_info}{A data frame containing colony information for all loggers.}

\item{filter_list}{A list of filter settings for different species. Defaults to `seatrack_settings_list`.}

\item{extra_metadata}{A data frame containing extra metadata for all loggers.}

\item{show_filter_plots}{A logical indicating whether to show filter plots. Defaults to FALSE.}

\item{export_maps}{A logical indicating whether to export maps of the processed positions. Defaults to TRUE.}

\item{plotting_dir}{An optional directory path to save plotting outputs. Defaults to NULL.}

\item{output_dir}{An optional directory path to save processed outputs. Defaults to NULL.}

\item{calibration_mode}{A logical indicating whether to run in calibration mode. Defaults to TRUE}

\item{split_years}{A character string indicating the month and day to split light data into time windows on when only deployment and retrieval dates are provided (format "MM-DD"). Defaults to "06-01".}

\item{analyzer}{An optional string indicating the analyzer who provided calibration data, if this column is not provided in calibration data. Defaults to an empty string.}
}
\value{
If calibration_mode is FALSE, returns a list containing:
         - `twilight_estimates`: A data frame of twilight estimates.
         - `posdata_export`: A data frame of processed position data.
         - `filtering`: A data frame summarizing the filtering steps applied.
If calibration_mode is TRUE, returns data frame of default calibration outputs and (if plottinging dir is not NULL) exports calibration plots.
}
\description{
Given a logger ID and year, this function processes all light data files for that logger/year combination,
applying calibration data, filter settings, colony info, and extra metadata as needed.
For more information on the exact steps see the help file for `process_logger_positions()`.
}
