% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_filter_file.R
\name{read_filter_file}
\alias{read_filter_file}
\title{Read Filter File}
\usage{
read_filter_file(filepath)
}
\arguments{
\item{filepath}{Filepath to the filter settings file (Excel or CSV).}
}
\value{
A GLSfilterList object containing the filter settings.
}
\description{
Read filter settings from an Excel or CSV file and return as a GLSfilterList object.
The filter file should contain columns for logger_id, species, colony, and various filter settings.
See `create_filter_file` for an example of the expected format.
}
