% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seatrack_settings_list}
\alias{seatrack_settings_list}
\title{Seatrack filter settings}
\format{
An object of class \code{GLSFilterSettingsList} (inherits from \code{R6}) of length 5.
}
\usage{
seatrack_settings_list
}
\description{
Default seatrack filter settings in as \code{GLSFilterSettingsList} object.
}
\concept{data}
\keyword{datasets}
